Remove dependency on the ffmpeg library
This avoids the need to link against ffmpeg when building av1an,
and instead uses the ffmpeg/ffprobe binaries in all places where we need
it. We were already using the binaries in the majority of places, so
this simply results in an improvement for setup for the end user,
as linking against ffmpeg was a particular pain point.