Remove parking_lot dependency
As of Rust 1.62 the `std::sync::Mutex` was refactored to make its
performance comparable to `parking_lot::Mutex`, which means there's no
need for us to import this additional dependency anymore.

This also found and fixed and issue where we were creating a lock twice
for every line of stderr parsing, which should reduce performance
overhead.