Use `VapoursynthDecoder` instead of `vspipe` for scene detection
There was a comment about `vspipe` being faster, however that was
incorrect from my testing. In fact, my testing showed that for a 1080p
H.264 input, `VapoursynthDecoder` was about 30% faster than `vspipe`,
likely due to avoiding the extra memory allocations that come from
needing to pipe the data. Therefore, we should be preferring avoiding
pipes where possible.

As a note, `FfmpegDecoder` is about 10% faster still compared to
`VapoursynthDecoder`. Piping `ffmpeg` to `Y4mDecoder` is about
equivalent to the `VapoursynthDecoder`. I may re-add the usage of
`FfmpegDecoder` as an *optional* feature in the future (since it
requires linking to the ffmpeg library), but will not do so in this PR.