fix: correct logic for limiting decoder threads
Decoder threads should only be limited to 1 in the case that we are
doing chunking for the real encode. It is incorrectly being limited to 1
in the case where we are doing SCD without downscaling or format
conversion. This commit fixes the logic around that, which provides a
significant speedup to scenechange detection.